Refactor the file cache lifetime management
Implement reference counting for file_lookups. This is necessary because
the potential delay of writer tasks to the point after the archive was
already closed still need to check the referenced memory in order to
detected such a situation. Furthermore the persistent meta and the
working set were packed together into the same memory allocation for
better cache locality. This in turn prevented the use of shared_ptr as
the destruction and construction within the same memory storage requires
synchronisation which cannot be properly implemented with shared_ptr.

Prepare the refactor of the free_block_index and archive_index file
logic into their own classes, so they can keep track of changes.