Work around MSVC compiler bug in _BitScanForward
Manually deduce whether the result of _BitScanForward should be used or
not by checking the given data value for 0. Even though the
documentation clearly states that _BitScanForward should return 0 only
if the input mask was zero, the generated ASM in Visual Studio 15.6
is just completely unfit to check anything related to a zero input mask.