Enhance cache purging with dispose logic
Implement an additional purge method which executes a dispose functor
after the victim page has been purged while still holding the page lock.
We use this functionality to synchronize the tree walking algorithm and
page erasure. Therefore the purge function acquires an exclusive lock
even if the victim page does not reside in the cache.