Implement CBOR AE box utils
Add a function for wrapping a binary CBOR item in a CBOR array which
also includes MAC and salt. After laying out the CBOR structure the
binary gets encrypted with a user provided key.

Add a function for decoding a CBOR box, i.e. retrieving all necessary
bits from a buffer for decrypting the contained binary.