Import cbor tuple/object head decoding utils
Deeppack provides codecs for classes with a descriptor which however
depends on class type NTTPs. Clang hasn't implemented class type NTTPs
and it is uncertain when this will happen. In the mean time we imported
the sources of some tuple/object head decoding utilities which do not
require class type NTTPs, but handle stuff like versioning etc.