Add a personalization area header
Reserve the second half of the static archive header for user data. This
allows users to store data required to open/decrypt the archive within
the archive file. A typical use case would be to store a salt for a
password KDF or an (asymetrically) encypted key.

This change is fully backwards compatible, because the second half of
the static archive header hasn't been in use and is simply zeroed on
archive creation.