fix: Properly size the index tree write-stream
The `chunked_output_stream` base class requires an account of the
remaining space excluding the first buffer, while the
`chunked_input_stream` wants an inclusive figure. These got mixed up
which can lead to overwriting subsequent index entries.

In the future we need to get rid of the legacy shims, but that is
riskier than fixing this rather critical bug with the known bad API.